Move network defaults out of extra.wlansi.
